refactor: extrair LoginPage e SplashScreen de main.dart para lib/pages/
- Move LoginPage + _LoginPageState para lib/pages/login_page.dart
- Move _SplashScreen + _SplashScreenState para lib/pages/splash_screen.dart
  (renomeia _SplashScreen para SplashScreen para ser pública)
- Remove imports não utilizados de main.dart (auth_config, feedly_auth, login_screen)
- Adiciona parâmetro key ao construtor de SplashScreen (melhores práticas)
- Reduz main.dart de 634 para 356 linhas
- Todos os testes passam, flutter analyze limpo em lib/
